+++ lib/active_attr/matchers/have_attribute_matcher.rb
@@ -1,10 +1,11 @@
 module ActiveAttr
   module Matchers
-    # Specify that a model should have an attribute matching the criteria
+    # Specify that a model should have an attribute matching the criteria. See
+    # {HaveAttributeMatcher}
     #
     # @example Person should have a name attribute
     #   describe Person do
-    #     it { should have_attribute(:name) }
+    #     it { should have_attribute(:first_name) }
     #   end
     #
     # @param [Symbol, String, #to_sym] attribute_name
@@ -17,18 +18,32 @@ module ActiveAttr
     end
 
     # Verifies that an ActiveAttr-based model has an attribute matching the
-    # given criteria
+    # given criteria. See {Matchers#have_attribute}
     #
     # @since 0.2.0
     class HaveAttributeMatcher
-      # @return [Symbol]
-      # @private
-      attr_reader :attribute_name
+      # Specify that the attribute should have the given default value
+      #
+      # @example Person's first name should default to John
+      #   describe Person do
+      #     it do
+      #       should have_attribute(:first_name).with_default_value_of("John")
+      #     end
+      #   end
+      #
+      # @param [Object]
+      #
+      # @since 0.5.0
+      def with_default_value_of(default_value)
+        @default_value_set = true
+        @default_value = default_value
+        self
+      end
 
       # @return [String] Description
       # @private
       def description
-        "have attribute named #{attribute_name}"
+        "have attribute named #{attribute_name}#{ " with a default value of #{default_value.inspect}" if @default_value_set}"
       end
 
       # @return [String] Failure message
@@ -42,13 +57,15 @@ module ActiveAttr
       def initialize(attribute_name)
         raise TypeError, "can't convert #{attribute_name.class} into Symbol" unless attribute_name.respond_to? :to_sym
         @attribute_name = attribute_name.to_sym
+        @default_value_set = false
       end
 
       # @private
       def matches?(model_or_model_class)
         @model_class = class_from(model_or_model_class)
+        @attribute_definition = @model_class.attributes[attribute_name]
 
-        !!@model_class.attributes[attribute_name]
+        !!(@attribute_definition && (!@default_value_set || @attribute_definition[:default] == default_value))
       end
 
       # @return [String] Negative failure message
@@ -58,6 +75,8 @@ module ActiveAttr
       end
 
       private
+
+      attr_reader :attribute_name, :default_value
 
       def class_from(object)
         Class === object ? object : object.class