+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # ActiveAttr 0.5.0 (unreleased) #
 
+* Added AttributeDefaults
+* Added AttributeDefinition#[]
 * Added AttributeDefinition#type
 * Added Attributes.attribute_names
+* Added Matchers::HaveAttributeMatcher#with_default_value_of
 * Added TypecastedAttributes
 * Added Typecasting
 * Changed Attributes.attributes return value from an Array to a Hash
 * Changed redefining an attribute to actually redefine the attribute
-* Removed StrictMassAssignment, use StrictMassAssignment with ActiveModel v3.2
-  which allows assigning mass_assignment_sanitizer to :strict on the class
+* Removed StrictMassAssignment, instead use MassAssignmentSecurity with
+  ActiveModel v3.2 which allows assigning mass_assignment_sanitizer to
+  :strict on the class
 
 # ActiveAttr 0.4.1 (November 27, 2011) #